myproject/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
elasticsearch
gunicorn
